+++ player2/components/include/ring_blocking.h
@@ -44,23 +44,23 @@ Date        Modification                                    Name
 
 class RingBlocking_c : public Ring_c
 {
-    private:
+private:
 
-        OS_Event_t           Signal;
-        unsigned int         Limit;
-        unsigned int         NextExtract;
-        unsigned int         NextInsert;
-        uintptr_t           *Storage;
+    OS_Event_t           Signal;
+    unsigned int         Limit;
+    unsigned int         NextExtract;
+    unsigned int         NextInsert;
+    unsigned int        *Storage;
 
-    public:
+public:
 
-        RingBlocking_c(unsigned int MaxEntries = 16);
-        ~RingBlocking_c(void);
+    RingBlocking_c( unsigned int MaxEntries = 16 );
+    ~RingBlocking_c( void );
 
-        RingStatus_t Insert(uintptr_t    Value);
-        RingStatus_t Extract(uintptr_t  *Value);
-        RingStatus_t Flush(void);
-        bool         NonEmpty(void);
+    RingStatus_t Insert( unsigned int    Value );
+    RingStatus_t Extract( unsigned int  *Value );
+    RingStatus_t Flush( void );
+    bool         NonEmpty( void );
 };
 
 #endif
