+++ player2/components/include/ring_unprotected.h
@@ -44,22 +44,22 @@ Date        Modification                                    Name
 
 class RingUnprotected_c : public Ring_c
 {
-    private:
+private:
 
-        unsigned int         Limit;
-        unsigned int         NextExtract;
-        unsigned int         NextInsert;
-        uintptr_t           *Storage;
+    unsigned int         Limit;
+    unsigned int         NextExtract;
+    unsigned int         NextInsert;
+    unsigned int        *Storage;
 
-    public:
+public:
 
-        RingUnprotected_c(unsigned int MaxEntries = 16);
-        ~RingUnprotected_c(void);
+    RingUnprotected_c( unsigned int MaxEntries = 16 );
+    ~RingUnprotected_c( void );
 
-        RingStatus_t Insert(uintptr_t    Value);
-        RingStatus_t Extract(uintptr_t  *Value);
-        RingStatus_t Flush(void);
-        bool         NonEmpty(void);
+    RingStatus_t Insert( unsigned int    Value );
+    RingStatus_t Extract( unsigned int  *Value );
+    RingStatus_t Flush( void );
+    bool         NonEmpty( void );
 };
 
 #endif
